+++ changelog.md
@@ -18,6 +18,7 @@
 - Offset Problem with Basic Revised \#2829
   - After installing the update, go to the settings for `GURPS 4e Game Aid > PDF Settings`, and change `Basic Set PDFs` to "Basic Revised ('B')".
 - Detached combat tracker window/popout! window not allowing updates to maneuvers. \#2796
+- GCS v5 import: spell points inside containers are not summed (+ minor: importSp never sets spl.level) \#2772
 
 ## Release 0.18.20 06/18/2026